use pytest-simcore improve fixtures and logs
