+++ tests/e2e-playwright/requirements/dev.txt
@@ -21,3 +21,4 @@
 --requirement _tools.txt
 
 # installs this repo's packages
+--editable ../../packages/pytest-simcore
